pyproject.toml
[tool.poetry]
name = "exif-frame-cli"
version = "0.1.0"
description = "A CLI tool for creating instant camera-style frames on digital photos using EXIF metadata"
authors = ["FScoward"]
readme = "README.md"
packages = [{include = "exif_frame_cli", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8.1"
click = "^8.1.0"
pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
flake8 = "^6.0.0"
pytest = "^7.0.0"

[tool.poetry.scripts]
exif-frame-cli = "exif_frame_cli.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]